Resolve all errors and add comprehensive game systems
- Fix compilation errors across all registry systems and components
- Add new gameplay systems: combat, mutiny, automation, diplomacy, economy
- Add comprehensive test coverage and test utilities
- Update documentation and error references
- Add dev menu and editor tools for scene setup
- Add mining scenario system and demo scenarios

+++ Assets/Scripts/Space4x/Authoring/Space4XMiningDemoAuthoring.cs
@@ -634,7 +634,7 @@ namespace Space4X.Registry
                     AddComponent(entity, new ResourceSourceConfig
                     {
                         GatherRatePerWorker = math.max(0.1f, asteroid.MiningRate),
-                        MaxSimultaneousWorkers = maxWorkers,
+                        MaxSimultaneousWorkers = (ushort)maxWorkers,
                         RespawnSeconds = 0f,
                         Flags = 0
                     });
